% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_pop_estimates.R
\name{extend_pop_estimates}
\alias{extend_pop_estimates}
\title{Extend population Estimates}
\usage{
extend_pop_estimates(pops_df, max_pop_year, year_var = "year")
}
\arguments{
\item{pops_df}{a \link[tibble:tibble-package]{tibble}
containing population estimates}

\item{max_pop_year}{The latest year required}

\item{year_var}{The variable containing the year}
}
\value{
a \link[tibble:tibble-package]{tibble} with extra years added
}
\description{
Take a \link[tibble:tibble-package]{tibble} containing population
estimates per year and extend this by copying the latest estimates \code{n}
times up to \code{max_pop_year}
}
