% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer_like.R
\name{is_integer_like}
\alias{is_integer_like}
\title{Check if an object looks like an integer}
\usage{
is_integer_like(x)
}
\arguments{
\item{x}{a vector of values to check}
}
\value{
TRUE or FALSE indicating if all values
look like integers - if the vector is not of type
character or numeric it will always return FALSE
}
\description{
Take a vector and check to see if
it looks like it contains only integers.
}
\examples{
library(dplyr)
data <-
  tibble(
    x = c(1, 2, 3),
    y = c("4", "5", "6"),
    z = c("a", "b", "c")
  )
data \%>\%
  mutate(across(where(is_integer_like), as.integer))
}
