% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_beddays_and_admissions.R
\name{monthly_beddays_and_admissions}
\alias{monthly_beddays_and_admissions}
\title{Generate counts of beddays per month for an episode
with an admission and discharge date
Optionally these months can be between two
supplied dates}
\usage{
monthly_beddays_and_admissions(
  data,
  admission_var,
  discharge_var,
  earliest_date = NA,
  latest_date = NA,
  pivot_longer = TRUE
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} with date variables \code{admission_date} and \code{discharge_date}}

\item{admission_var, discharge_var}{the \code{admission_date} and \code{discharge_date} variables}

\item{earliest_date, latest_date}{the earliest and latest dates (months) to compute new variables for}

\item{pivot_longer}{Should the data be left in wide form or pivotted to long form (default)}
}
\value{
a \link[tibble:tibble-package]{tibble} with new variables
}
\description{
Generate counts of beddays per month for an episode
with an admission and discharge date
Optionally these months can be between two
supplied dates
}
