% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_database_conn.R
\name{phs_database_conn}
\alias{phs_database_conn}
\title{Open a connection to a PHS database}
\usage{
phs_database_conn(dsn, username = Sys.getenv("USER"))
}
\arguments{
\item{dsn}{The Data Source Name passed on to \code{odbc::dbconnect}
the dsn must be setup first. e.g. SMRA or DVPROD}

\item{username}{The username to use for authentication,
if not supplied it first will check the environment variable
and finally ask the user for input.}
}
\value{
a connection to the specified dsn
}
\description{
Open a connection to a PHS database
}
